Improved routes

+++ src/com/karcz/piotr/routes/CartRoute.kt
@@ -3,6 +3,9 @@ package com.karcz.piotr.routes
 import com.karcz.piotr.data.CartModel
 import com.karcz.piotr.repository.dao.CartDao
 import com.karcz.piotr.repository.dao.CartDaoImpl
+import com.karcz.piotr.repository.dao.ProductDao
+import com.karcz.piotr.repository.dao.ProductDaoImpl
+import com.karcz.piotr.transfer.data.CartResponseModel
 import com.karcz.piotr.transfer.data.Response
 import io.ktor.application.*
 import io.ktor.auth.*
@@ -14,6 +17,7 @@ import java.sql.SQLException
 
 fun Route.cartRoute() {
     val cartDao: CartDao = CartDaoImpl()
+    val productDao: ProductDao = ProductDaoImpl()
 
     authenticate {
         route("/me/cart") {
@@ -21,7 +25,8 @@ fun Route.cartRoute() {
                 val email = call.principal<UserIdPrincipal>()?.name
                 if (email != null) {
                     val cartItems: List<CartModel> = cartDao.getAllForClient(email)
-                    call.respond(HttpStatusCode.OK, cartItems)
+                    val productItems = cartItems.mapNotNull { productDao.get(it.productId) }
+                    call.respond(HttpStatusCode.OK, CartResponseModel(cartItems, productItems))
                 } else {
                     call.respond(HttpStatusCode.NotFound)
                 }
@@ -32,7 +37,7 @@ fun Route.cartRoute() {
             post {
                 val cartModel = try {
                     call.receive<CartModel>()
-                } catch (e: ContentTransformationException) {
+                } catch (e: Exception) {
                     call.respond(HttpStatusCode.BadRequest)
                     return@post
                 }
@@ -44,11 +49,19 @@ fun Route.cartRoute() {
                 }
 
                 try {
-                    if (cartDao.isIn(cartModel)) {
-                        val currentCartItem = cartDao.get(cartModel.customerEmail, cartModel.productId)
-                        cartDao.update(cartModel = cartModel.copy(quantity = cartModel.quantity + (currentCartItem?.quantity ?: 0)))
+                    if (productDao.isAvailable(cartModel.productId, cartModel.quantity)) {
+                        if (cartDao.isIn(cartModel)) {
+                            val currentCartItem = cartDao.get(cartModel.customerEmail, cartModel.productId)
+                            cartDao.update(
+                                cartModel = cartModel.copy(
+                                    quantity = cartModel.quantity + (currentCartItem?.quantity ?: 0)
+                                )
+                            )
+                        } else {
+                            cartDao.add(cartModel)
+                        }
                     } else {
-                        cartDao.add(cartModel)
+                        call.respond(HttpStatusCode.OK, Response(false, "There are less available products of this id."))
                     }
                 } catch (e: SQLException) {
                     call.respond(HttpStatusCode.InternalServerError)
@@ -62,7 +75,7 @@ fun Route.cartRoute() {
             delete {
                 val cartModel = try {
                     call.receive<CartModel>()
-                } catch (e: ContentTransformationException) {
+                } catch (e: Exception) {
                     call.respond(HttpStatusCode.BadRequest)
                     return@delete
                 }
@@ -90,7 +103,7 @@ fun Route.cartRoute() {
             put {
                 val cartModel = try {
                     call.receive<CartModel>()
-                } catch (e: ContentTransformationException) {
+                } catch (e: Exception) {
                     call.respond(HttpStatusCode.BadRequest)
                     return@put
                 }
@@ -101,19 +114,23 @@ fun Route.cartRoute() {
                     return@put
                 }
 
-                if (cartDao.isIn(cartModel)) {
-                    try {
-                        if (cartModel.quantity > 0) {
-                            cartDao.update(cartModel)
-                        } else {
-                            cartDao.remove(cartModel)
+                if (productDao.isAvailable(cartModel.productId, cartModel.quantity)) {
+                    if (cartDao.isIn(cartModel)) {
+                        try {
+                            if (cartModel.quantity > 0) {
+                                cartDao.update(cartModel)
+                            } else {
+                                cartDao.remove(cartModel)
+                            }
+                        } catch (e: SQLException) {
+                            call.respond(HttpStatusCode.InternalServerError)
                         }
-                    } catch (e: SQLException) {
-                        call.respond(HttpStatusCode.InternalServerError)
+                        call.respond(HttpStatusCode.OK, Response(true, "Item updated."))
+                    } else {
+                        call.respond(HttpStatusCode.OK, Response(false, "Item is not in cart."))
                     }
-                    call.respond(HttpStatusCode.OK, Response(true, "Item updated."))
                 } else {
-                    call.respond(HttpStatusCode.OK, Response(false, "Item is not in cart."))
+                    call.respond(HttpStatusCode.OK, Response(false, "There are less available products of this id."))
                 }
             }
         }
